Refactoring

+++ training/DataStructures-Algorithm/sorting/mergeSort.py
@@ -1,6 +1,6 @@
 def merge_sort(arr):
     if len(arr)>1:
-        mid = len(arr) / 2
+        mid = int(len(arr) / 2)
         L = arr[:mid]
         R = arr[mid:]
         merge_sort(L)
